# openapi.yaml
openapi: 3.1.0
info:
  title: MedQuiz API
  version: 1.0.0
  description: |
    API contract v1 for MedQuiz (PASS/LAS -> DFGSM/DFASM -> Interne/Docteur junior),
    aligned with Lots B/C/D/E.
servers:
  - url: https://api.medquiz.app
    description: Production
  - url: https://staging-api.medquiz.app
    description: Staging
security:
  - bearerAuth: []
...
paths:
  /v1/auth/register:
    post:
      tags: [Auth]
      security: []
      summary: Register account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: Registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSuccessResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
  /v1/auth/login:
    post:
      tags: [Auth]
      security: []
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Logged in
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSuccessResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /v1/auth/refresh:
    post:
      tags: [Auth]
      security: []
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSuccessResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
  /v1/duels:
    post:
      tags: [Duel]
      summary: Create duel
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDuelRequest'
      responses:
        '201':
          description: Duel created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateDuelResponse'
        '422':
          $ref: '#/components/responses/BusinessRuleError'
    get:
      tags: [Duel]
      summary: List duels
      parameters:
        - $ref: '#/components/parameters/StatusFilter'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Duel list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DuelListResponse'
  /v1/duels/{duelId}:
    get:
      tags: [Duel]
      summary: Get duel details
      parameters:
        - $ref: '#/components/parameters/DuelId'
      responses:
        '200':
          description: Duel details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DuelDetailResponse'
  /v1/duels/{duelId}/accept:
    post:
      tags: [Duel]
      summary: Accept duel invitation
      parameters:
        - $ref: '#/components/parameters/DuelId'
      responses:
        '200':
          $ref: '#/components/responses/OkEnvelope'
  /v1/duels/{duelId}/decline:
    post:
      tags: [Duel]
      summary: Decline duel invitation
      parameters:
        - $ref: '#/components/parameters/DuelId'
      responses:
        '200':
          $ref: '#/components/responses/OkEnvelope'
  /v1/duels/{duelId}/opener:
    get:
      tags: [Duel]
      summary: Get opener question
      parameters:
        - $ref: '#/components/parameters/DuelId'
      responses:
        '200':
          description: Opener question
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpenerQuestionResponse'
  /v1/duels/{duelId}/opener/answer:
    post:
      tags: [Duel]
      summary: Submit opener answer
      parameters:
        - $ref: '#/components/parameters/DuelId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OpenerAnswerRequest'
      responses:
        '200':
          $ref: '#/components/responses/OkEnvelope'
  /v1/duels/{duelId}/opener/decision:
    post:
      tags: [Duel]
      summary: Winner chooses take or leave hand
      parameters:
        - $ref: '#/components/parameters/DuelId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OpenerDecisionRequest'
      responses:
        '200':
          $ref: '#/components/responses/OkEnvelope'
  /v1/duels/{duelId}/rounds/current:
    get:
      tags: [Duel]
      summary: Get current round state
      parameters:
        - $ref: '#/components/parameters/DuelId'
      responses:
        '200':
          description: Round state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentRoundResponse'
  /v1/duels/{duelId}/rounds/{roundNo}/choose-subject:
    post:
      tags: [Duel]
      summary: Choose subject for current round
      parameters:
        - $ref: '#/components/parameters/DuelId'
        - $ref: '#/components/parameters/RoundNo'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChooseSubjectRequest'
      responses:
        '200':
          $ref: '#/components/responses/OkEnvelope'
        '422':
          $ref: '#/components/responses/BusinessRuleError'
  /v1/duels/{duelId}/rounds/{roundNo}/questions:
    get:
      tags: [Duel]
      summary: Get assigned round questions for current player
      parameters:
        - $ref: '#/components/parameters/DuelId'
        - $ref: '#/components/parameters/RoundNo'
      responses:
        '200':
          description: Question list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DuelRoundQuestionsResponse'
  /v1/duels/{duelId}/rounds/{roundNo}/answers:
    post:
      tags: [Duel]
      summary: Submit one round answer (unit mode)
      parameters:
        - $ref: '#/components/parameters/DuelId'
        - $ref: '#/components/parameters/RoundNo'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DuelRoundAnswerRequest'
      responses:
        '200':
          description: Round answer result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DuelRoundAnswerResponse'
        '422':
          $ref: '#/components/responses/BusinessRuleError'
  /v1/duels/{duelId}/jokers/request:
    post:
      tags: [Duel]
      summary: Request +24h joker
      parameters:
        - $ref: '#/components/parameters/DuelId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JokerRequest'
      responses:
        '200':
          $ref: '#/components/responses/OkEnvelope'
        '422':
          $ref: '#/components/responses/BusinessRuleError'
  /v1/duels/{duelId}/jokers/{jokerId}/respond:
    post:
      tags: [Duel]
      summary: Respond to joker request
      parameters:
        - $ref: '#/components/parameters/DuelId'
        - $ref: '#/components/parameters/JokerId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JokerResponseRequest'
      responses:
        '200':
          $ref: '#/components/responses/OkEnvelope'
  /v1/duels/{duelId}/forfeit:
    post:
      tags: [Duel]
      summary: Forfeit duel
      parameters:
        - $ref: '#/components/parameters/DuelId'
      responses:
        '200':
          $ref: '#/components/responses/OkEnvelope'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
...
    DuelId:
      name: duelId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    RoundNo:
      name: roundNo
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
        maximum: 5
    JokerId:
      name: jokerId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Cursor:
      name: cursor
      in: query
      required: false
      schema:
        type: string
    StatusFilter:
      name: status
      in: query
      required: false
      schema:
        type: string
  responses:
    OkEnvelope:
      description: Generic ok response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/OkResponse'
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnauthorizedError:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    BusinessRuleError:
      description: Business rule violation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    DuelStatus:
      type: string
      enum: [pending_opener, in_progress, completed, cancelled, expired]
...
    OkResponse:
      type: object
      properties:
        data:
          type: object
          additionalProperties: true
          default: {}
        meta:
          type: object
          additionalProperties: true
          default: {}
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: array
              items:
                type: object
                additionalProperties: true
            requestId:
              type: string
    TokenPair:
      type: object
      required: [accessToken, refreshToken]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
    UserPublic:
      type: object
      required: [id, email, displayName]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        displayName:
          type: string
    AuthSuccessResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            user:
              $ref: '#/components/schemas/UserPublic'
            tokens:
              $ref: '#/components/schemas/TokenPair'
    RegisterRequest:
      type: object
      required: [email, password, displayName]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        displayName:
          type: string
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    RefreshRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
...
    Choice:
      type: object
      properties:
        id:
          type: string
          format: uuid
        label:
          type: string
        position:
          type: integer
          minimum: 1
          maximum: 4
    Question:
      type: object
      properties:
        id:
          type: string
          format: uuid
        prompt:
          type: string
        choices:
          type: array
          items:
            $ref: '#/components/schemas/Choice'
        subjectId:
          type: string
          format: uuid
        chapterId:
          type: string
          format: uuid
...
    CreateDuelRequest:
      type: object
      required: [matchmakingMode]
      properties:
        matchmakingMode:
          type: string
          enum: [friend_invite, random_free, random_level]
        opponentUserId:
          type: string
          format: uuid
          nullable: true
    CreateDuelResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            duelId:
              type: string
              format: uuid
            status:
              $ref: '#/components/schemas/DuelStatus'
    DuelListResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            items:
              type: array
              items:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  status:
                    $ref: '#/components/schemas/DuelStatus'
                  player1Score:
                    type: integer
                  player2Score:
                    type: integer
            meta:
              type: object
              properties:
                nextCursor:
                  type: string
                  nullable: true
    DuelDetailResponse:
      type: object
      properties:
        data:
          type: object
          additionalProperties: true
    OpenerQuestionResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            question:
              $ref: '#/components/schemas/Question'
    OpenerAnswerRequest:
      type: object
      required: [selectedChoiceId, responseTimeMs]
      properties:
        selectedChoiceId:
          type: string
          format: uuid
        responseTimeMs:
          type: integer
          minimum: 1
    OpenerDecisionRequest:
      type: object
      required: [decision]
      properties:
        decision:
          type: string
          enum: [take_hand, leave_hand]
    CurrentRoundResponse:
      type: object
      properties:
        data:
          type: object
          additionalProperties: true
    ChooseSubjectRequest:
      type: object
      required: [subjectId]
      properties:
        subjectId:
          type: string
          format: uuid
    DuelRoundQuestionsResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/Question'
    DuelRoundAnswerRequest:
      type: object
      required: [slotNo, questionId, selectedChoiceId, responseTimeMs]
      properties:
        slotNo:
          type: integer
          minimum: 1
          maximum: 3
        questionId:
          type: string
          format: uuid
        selectedChoiceId:
          type: string
          format: uuid
        responseTimeMs:
          type: integer
          minimum: 1
    DuelRoundAnswerResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            answerResult:
              type: object
              properties:
                isCorrect:
                  type: boolean
                correctChoiceId:
                  type: string
                  format: uuid
                explanation:
                  type: string
            roundProgress:
              type: object
              properties:
                answeredSlots:
                  type: integer
                remainingSlots:
                  type: integer
            turnCompleted:
              type: boolean
    JokerRequest:
      type: object
      properties:
        reason:
          type: string
    JokerResponseRequest:
      type: object
      required: [decision]
      properties:
        decision:
          type: string
          enum: [grant, reject]
...